requirements.txt
pandas==2.2.2
# scispacy==0.5.4
beautifulsoup4==4.12.3
rank_bm25==0.2.2
pymongo==4.8.0
python-dotenv==1.0.0
sentence_transformers==3.0.1
torch==2.3.1
flair==0.14.0
diskcache
SPARQLWrapper==2.0.0
openai==1.47.0
icd10-cm==0.0.5
googletrans==3.1.0a0
huggingface_hub==0.24.6
transformers==4.45.1
ollama==0.3.12
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1
seaborn==0.13.2
